+++ docs/src/app/(docs)/layout.tsx
@@ -1,12 +1,14 @@
 import "@/app/(docs)/styles.css";
 import { baseOptions } from "@/app/layout.config";
 import HeaderBar from "@/components/header-bar";
+import TamboHackBanner from "@/components/tambohack-banner";
 import { source } from "@/lib/source";
 import { DocsLayout } from "fumadocs-ui/layouts/docs";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <div>
+      <TamboHackBanner />
       <HeaderBar />
       <DocsLayout
         tree={source.pageTree}